requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
markdownify>=0.13.1
pypdf>=4.3.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
openai>=1.40.0

